requirements.txt
gym==0.26.2
gym_gridverse @ git+https://github.com/josndan/gym-gridverse.git@master
numpy~=1.26.4
pyglet~=1.5.27
matplotlib~=3.9.2
omegaconf~=2.3.0
clearml~=1.16.3
torch
torchvision
torchaudio
tensorboard
hydra-core
pillow~=10.4.0
moviepy~=1.0.3